Create 10_00M_demo
this will give a "OOM error". This is because the “OOM-demo” pod runs a process that consumes 15Mi of memory. Increase the limit of the pod to 20Mi.
kubectl get po OOM-demo -o yaml > oom.yaml and edit the file such as memory limit is set to 20Mi . Then run kubectl replace -f oom.yaml --force. This command will delete the existing one first and recreate a new one from the YAML file.